panel: nt37290: force to disable LHBM while disabling panel
NT37290 has the constraint that we cannot turn off the panel
directly while LHBM is already enabled, otherwise a green flicker
will appear obviously. We have to disable LHBM first so that the
panel can be disabled. This change forces to disable LHBM
immediately in this situation to avoid the problem.

Bug: 258705400
Test: turn off display while LHBM is enabled
Change-Id: I0702ecc8332505fb96a13bc43abc6aa49dce86fc
Signed-off-by: Chris Lu <luchris@google.com>